SCRD-2448 Update in-house certifi version
certifi is now required by the Python requests library, which
means that installing requests is currently pulling in an
upstream version of certifi that does not load the SSL
certs from the system location. Modify the version so that
the certifi we install from source appears newer than the
upstream. This should help with the many SSL certificate
related issues we are experiencing right now in QE testing.

Change-Id: I3859910aecaefe9b0fe6bf23e83b1960fd501b6b